+++ src/Chunkyard.Tests/Core/DataConvertTests.cs
@@ -3,54 +3,9 @@ namespace Chunkyard.Tests.Core;
 public static class DataConvertTests
 {
     [Fact]
-    public static void DataConvert_Converts_Objects_To_Bytes_And_Back()
+    public static void DataConvert_Can_Convert_Snapshot()
     {
-        var blob = new Blob("some blob", DateTime.UtcNow);
-
-        var bytes = DataConvert.ObjectToBytes(blob);
-
-        Assert.Equal(
-            blob,
-            DataConvert.BytesToObject<Blob>(bytes));
-    }
-
-    [Fact]
-    public static void DataConvert_Converts_Text_To_Bytes_And_Back()
-    {
-        var text = "Hello World!";
-
-        var bytes = DataConvert.TextToBytes(text);
-
-        Assert.Equal(
-            text,
-            DataConvert.BytesToText(bytes));
-    }
-
-    [Fact]
-    public static void DataConvert_Can_Read_Serialized_Snapshot()
-    {
-        var serializedSnapshot = new
-        {
-            CreationTimeUtc = "2020-05-07T18:33:00Z",
-            BlobReferences = new[]
-            {
-                new
-                {
-                    Blob = new
-                    {
-                        Name = "some blob",
-                        LastWriteTimeUtc = "2020-05-07T18:33:00Z"
-                    },
-                    Nonce = "ESIzRA==",
-                    ChunkIds = new[]
-                    {
-                        "sha256://ad95131bc0b799c0b1af477fb14fcf26a6a9f76079e48bf090acb7e8367bfd0e"
-                    }
-                }
-            }
-        };
-
-        var currentSnapshot = new Snapshot(
+        var expected = new Snapshot(
             new DateTime(2020, 05, 07, 18, 33, 0, DateTimeKind.Utc),
             new[]
             {
@@ -65,33 +20,16 @@ public static class DataConvertTests
                     })
             });
 
-        var expected = DataConvert.BytesToText(
-            DataConvert.ObjectToBytes(serializedSnapshot));
+        var actual = DataConvert.BytesToSnapshot(
+            DataConvert.SnapshotToBytes(expected));
 
-        var actual = DataConvert.BytesToText(
-            DataConvert.ObjectToBytes(currentSnapshot));
-
-        Assert.True(
-            expected.Equals(actual),
-            "Broke persisted repository format");
+        Assert.Equal(expected, actual);
     }
 
     [Fact]
-    public static void DataConvert_Can_Read_Serialized_SnapshotReference()
+    public static void DataConvert_Can_Convert_SnapshotReference()
     {
-        var serializedReference = new
-        {
-            SchemaVersion = 1,
-            Salt = "ESIzRA==",
-            Iterations = 1000,
-            ChunkIds = new[]
-            {
-                "sha256://ad95131bc0b799c0b1af477fb14fcf26a6a9f76079e48bf090acb7e8367bfd0e"
-            }
-        };
-
-        var currentReference = new SnapshotReference(
-            SnapshotStore.SchemaVersion,
+        var expected = new SnapshotReference(
             new byte[] { 0x11, 0x22, 0x33, 0x44 },
             1000,
             new[]
@@ -99,37 +37,9 @@ public static class DataConvertTests
                 new Uri("sha256://ad95131bc0b799c0b1af477fb14fcf26a6a9f76079e48bf090acb7e8367bfd0e")
             });
 
-        var expected = DataConvert.BytesToText(
-            DataConvert.ObjectToBytes(serializedReference));
-
-        var actual = DataConvert.BytesToText(
-            DataConvert.ObjectToBytes(currentReference));
-
-        Assert.True(
-            expected.Equals(actual),
-            "Broke persisted repository format");
-    }
-
-    [Fact]
-    public static void BytesToVersionedObject_Returns_Supported_Object()
-    {
-        var expected = new Versioned(1);
-
-        var actual = DataConvert.BytesToVersionedObject<Versioned>(
-            DataConvert.ObjectToBytes(expected),
-            1);
+        var actual = DataConvert.BytesToSnapshotReference(
+            DataConvert.SnapshotReferenceToBytes(expected));
 
         Assert.Equal(expected, actual);
-    }
-
-    [Fact]
-    public static void BytesToVersionedObject_Throws_If_Unsupported_Object()
-    {
-        var obj = new Versioned(1);
-
-        Assert.Throws<NotSupportedException>(
-            () => DataConvert.BytesToVersionedObject<Versioned>(
-                DataConvert.ObjectToBytes(obj),
-                2));
     }
 }