+++ src/Chunkyard/Core/JsonContext.cs
@@ -1,0 +1,9 @@
+using System.Text.Json.Serialization;
+
+namespace Chunkyard.Core;
+
+[JsonSerializable(typeof(Snapshot))]
+[JsonSerializable(typeof(SnapshotReference))]
+internal partial class JsonContext : JsonSerializerContext
+{
+}