+++ src/Chunkyard/Core/SnapshotReference.cs
@@ -4,21 +4,17 @@ namespace Chunkyard.Core;
 /// A reference which can be used to retrieve a <see cref="Snapshot"/> from
 /// a <see cref="SnapshotStore"/> based on a password based encryption key.
 /// </summary>
-public class SnapshotReference : IVersioned
+public class SnapshotReference
 {
     public SnapshotReference(
-        int schemaVersion,
         byte[] salt,
         int iterations,
         IReadOnlyCollection<Uri> chunkIds)
     {
-        SchemaVersion = schemaVersion;
         Salt = salt;
         Iterations = iterations;
         ChunkIds = chunkIds;
     }
-
-    public int SchemaVersion { get; }
 
     public byte[] Salt { get; }
 
@@ -29,7 +25,6 @@ public class SnapshotReference : IVersioned
     public override bool Equals(object? obj)
     {
         return obj is SnapshotReference other
-            && SchemaVersion == other.SchemaVersion
             && Salt.SequenceEqual(other.Salt)
             && Iterations == other.Iterations
             && ChunkIds.SequenceEqual(other.ChunkIds);
@@ -37,6 +32,6 @@ public class SnapshotReference : IVersioned
 
     public override int GetHashCode()
     {
-        return HashCode.Combine(SchemaVersion, Salt, Iterations, ChunkIds);
+        return HashCode.Combine(Salt, Iterations, ChunkIds);
     }
 }